+++ packages/bridge-schedule/src/index.ts
@@ -3,4 +3,5 @@ export { SUPPORTED_BRIDGES_NAME_SET } from "./const";
 export { getBridgeEvents } from "./events";
 export { getNavigationState, isNavigationTime } from "./navigation";
 export { getBridgeScheduleEntry, getBridgesState, getBridgeState, getNextBridgeEvent } from "./provider";
+export type { NavigationState } from "./types";
 export { isAllBridgesLiftedDown, isBridgeException } from "./utils";